+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- Added ViewModel.buildViewCalls to facilitate unit testing of ViewModels.
+- fixed bug where `name` was ignored in `ViewModel.listenTo(name: name)`.
+
 ## 1.0.6
 Edited the readme page.